ai-service/requirements.txt
# Smart Cooking AI Service - Enhanced Dependencies với Type Stubs
# Main FastAPI framework
fastapi==0.116.1
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# HTTP client for external API calls
httpx==0.28.1
requests==2.32.3

# Environment variables
python-dotenv==1.1.1

# Data validation and serialization
pydantic==2.10.3

# Google AI and APIs
google-generativeai==0.8.3
google-cloud-translate==3.12.1
google-cloud-vision==3.8.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Type checking và stubs cho Pylance
mypy==1.14.0
types-requests==2.32.0.20241016
types-python-dateutil==2.9.0.20241206

# Development tools
black==24.10.0
isort==5.13.2
pytest==8.3.4
pytest-asyncio==0.25.0
google-api-python-client==2.177.0

# OpenAI API
openai==1.3.7

# Image processing
Pillow>=10.0.0

# JSON handling and utilities
orjson==3.9.10

# Async support
asyncio-throttle==1.0.2

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security
python-multipart==0.0.6

# CORS support (included in FastAPI)
# Database support (if needed later)
sqlalchemy==2.0.23
databases[postgresql]==0.8.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
